refactor: change name folder homePage as dashboardAdmin,compose filterEvents
