+++ pages/admin/index.jsx
@@ -1,13 +1,13 @@
 import { Box } from "@chakra-ui/react";
 import { NavBar } from "../../components/admin/navbar/NavBar";
-import { HomePage } from "../../components/admin/listEvents/HomePage";
+import { DashboardAdmin } from "../../components/admin/listEvents/DashboardAdmin";
 
 const index = () => {
   return (
     <>
       <Box>
         <NavBar />
-        <HomePage />
+        <DashboardAdmin />
       </Box>
     </>
   );
